pyproject.toml
[project]
name = "supermetrics-cli-scripts"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = [
    "pyyaml~=6.0.3",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN",     # type annotations — not adding to existing scripts
    "D",       # docstrings — scripts, not a library
    "PLR2004", # magic values
    "PLR0912", # too many branches — generator is inherently complex
    "PLR0915", # too many statements — same reason
    "C901",    # complexity — same reason
    "EM101",   # string literals in exceptions
    "EM102",   # f-string literals in exceptions
    "TRY003",  # long exception messages
    "COM812",  # trailing comma (conflicts with formatter)
    "T201",    # print statements — generator uses print for stderr output
    "S101",    # assert — used in tests
    "E501",    # line too long — Go code generation has long string literals
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["ANN", "ARG001", "S101", "S105", "S106", "PT009", "PT027"]

[tool.ruff.lint.isort]
force-single-line = true
